+++ src/components/InngestCommHandler.ts
@@ -1,11 +1,13 @@
 import canonicalize from "canonicalize";
 import { hmac, sha256 } from "hash.js";
 import { z } from "zod";
+import { ServerTiming } from "../helpers/ServerTiming";
 import { envKeys, headerKeys, queryKeys } from "../helpers/consts";
 import { devServerAvailable, devServerUrl } from "../helpers/devserver";
+import { allProcessEnv, isProd } from "../helpers/env";
 import { serializeError } from "../helpers/errors";
 import { strBoolean } from "../helpers/scalar";
-import { ServerTiming } from "../helpers/ServerTiming";
+import { stringifyUnknown } from "../helpers/strings";
 import type { MaybePromise } from "../helpers/types";
 import { landing } from "../landing";
 import {
@@ -53,19 +55,33 @@ export type ServeHandler = (
    * The name of this app, used to scope and group Inngest functions, or
    * the `Inngest` instance used to declare all functions.
    */
-  nameOrInngest: string | Inngest,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  nameOrInngest: string | Inngest<any>,
 
   /**
    * An array of the functions to serve and register with Inngest.
    */
-  functions: InngestFunction[],
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  functions: InngestFunction<any, any, any>[],
 
   /**
    * A set of options to further configure the registration of Inngest
    * functions.
    */
   opts?: RegisterOptions
-) => unknown;
+  /**
+   * This `any` return is appropriate.
+   *
+   * While we can infer the signature of the returned value, we cannot guarantee
+   * that we have used the same types as the framework we are integrating with,
+   * which sometimes can cause frustrating collisions for a user that result in
+   * `as unknown as X` casts.
+   *
+   * Instead, we will use `any` here and have the user be able to place it
+   * anywhere they need.
+   */
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+) => any;
 
 /**
  * Capturing the global type of fetch so that we can reliably access it below.
@@ -227,7 +243,7 @@ export class InngestCommHandler<H extends Handler, TransformedRes> {
    * A private collection of just Inngest functions, as they have been passed
    * when instantiating the class.
    */
-  private readonly rawFns: InngestFunction[];
+  private readonly rawFns: InngestFunction<any, any, any>[];
 
   /**
    * A private collection of functions that are being served. This map is used
@@ -260,12 +276,14 @@ export class InngestCommHandler<H extends Handler, TransformedRes> {
      * receiving events from the same service, as you can reuse a single
      * definition of Inngest.
      */
-    appNameOrInngest: string | Inngest,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    appNameOrInngest: string | Inngest<any>,
 
     /**
      * An array of the functions to serve and register with Inngest.
      */
-    functions: InngestFunction[],
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    functions: InngestFunction<any, any, any>[],
     {
       inngestRegisterUrl,
       fetch,
@@ -478,12 +496,14 @@ export class InngestCommHandler<H extends Handler, TransformedRes> {
       "Server-Timing": timer.getHeader(),
     });
 
+    const env = actions.env ?? allProcessEnv();
+    this._isProd = actions.isProduction ?? isProd(env);
+
     try {
       const runRes = await actions.run();
 
       if (runRes) {
-        this._isProd = runRes.isProduction;
-        this.upsertSigningKeyFromEnv(runRes.env);
+        this.upsertSigningKeyFromEnv(env);
         this.validateSignature(runRes.signature, runRes.data);
 
         const stepRes = await this.runStep(
@@ -523,11 +543,10 @@ export class InngestCommHandler<H extends Handler, TransformedRes> {
 
       const viewRes = await actions.view();
       if (viewRes) {
-        this._isProd = viewRes.isProduction;
-        this.upsertSigningKeyFromEnv(viewRes.env);
+        this.upsertSigningKeyFromEnv(env);
 
         const showLandingPage = this.shouldShowLandingPage(
-          viewRes.env[envKeys.LandingPage]
+          stringifyUnknown(env[envKeys.LandingPage])
         );
 
         if (this._isProd || !showLandingPage) {
@@ -540,8 +559,10 @@ export class InngestCommHandler<H extends Handler, TransformedRes> {
 
         if (viewRes.isIntrospection) {
           const introspection: IntrospectRequest = {
-            ...this.registerBody(this.reqUrl(viewRes.url)),
-            devServerURL: devServerUrl(viewRes.env[envKeys.DevServerUrl]).href,
+            ...this.registerBody(this.reqUrl(actions.url)),
+            devServerURL: devServerUrl(
+              stringifyUnknown(env[envKeys.DevServerUrl])
+            ).href,
             hasSigningKey: Boolean(this.signingKey),
           };
 
@@ -567,12 +588,11 @@ export class InngestCommHandler<H extends Handler, TransformedRes> {
 
       const registerRes = await actions.register();
       if (registerRes) {
-        this._isProd = registerRes.isProduction;
-        this.upsertSigningKeyFromEnv(registerRes.env);
+        this.upsertSigningKeyFromEnv(env);
 
         const { status, message } = await this.register(
-          this.reqUrl(registerRes.url),
-          registerRes.env[envKeys.DevServerUrl],
+          this.reqUrl(actions.url),
+          stringifyUnknown(env[envKeys.DevServerUrl]),
           registerRes.deployId
         );
 
@@ -877,9 +897,9 @@ export class InngestCommHandler<H extends Handler, TransformedRes> {
     return this._isProd;
   }
 
-  private upsertSigningKeyFromEnv(env: Record<string, string | undefined>) {
+  private upsertSigningKeyFromEnv(env: Record<string, unknown>) {
     if (!this.signingKey && env[envKeys.SigningKey]) {
-      this.signingKey = env[envKeys.SigningKey];
+      this.signingKey = String(env[envKeys.SigningKey]);
     }
   }
 
@@ -1022,6 +1042,10 @@ class RequestSignature {
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 type Handler = (...args: any[]) => {
+  env?: Record<string, unknown>;
+  isProduction?: boolean;
+  url: URL;
+} & {
   [K in Extract<
     HandlerAction,
     { action: "run" | "register" | "view" }
@@ -1059,22 +1083,13 @@ type HandlerAction =
   | {
       action: "error";
       data: Record<string, string>;
-      env: Record<string, string | undefined>;
-      isProduction: boolean;
-      url: URL;
     }
   | {
       action: "view";
-      env: Record<string, string | undefined>;
-      url: URL;
       isIntrospection: boolean;
-      isProduction: boolean;
     }
   | {
       action: "register";
-      env: Record<string, string | undefined>;
-      url: URL;
-      isProduction: boolean;
       deployId?: null | string;
     }
   | {
@@ -1082,14 +1097,8 @@ type HandlerAction =
       fnId: string;
       stepId: string | null;
       data: Record<string, unknown>;
-      env: Record<string, string | undefined>;
-      isProduction: boolean;
-      url: URL;
       signature: string | undefined;
     }
   | {
       action: "bad-method";
-      env: Record<string, string | undefined>;
-      isProduction: boolean;
-      url: URL;
     };