+++ src/express.ts
@@ -4,7 +4,6 @@ import {
   ServeHandler,
 } from "./components/InngestCommHandler";
 import { headerKeys, queryKeys } from "./helpers/consts";
-import { allProcessEnv } from "./helpers/env";
 
 /**
  * Serve and register any declared functions with Inngest, making them available
@@ -22,20 +21,15 @@ export const serve: ServeHandler = (nameOrInngest, fns, opts) => {
       const hostname = req.get("host") || req.headers["host"];
       const protocol = hostname?.includes("://") ? "" : `${req.protocol}://`;
       const url = new URL(req.originalUrl, `${protocol}${hostname || ""}`);
-      const env = allProcessEnv();
-      const isProduction =
-        env.ENVIRONMENT === "production" || env.NODE_ENV === "production";
 
       return {
+        url,
         run: () => {
           if (req.method === "POST") {
             return {
               fnId: req.query[queryKeys.FnId] as string,
               stepId: req.query[queryKeys.StepId] as string,
               data: req.body as Record<string, unknown>,
-              env,
-              isProduction,
-              url,
               signature: req.headers[headerKeys.Signature] as string,
             };
           }
@@ -43,9 +37,6 @@ export const serve: ServeHandler = (nameOrInngest, fns, opts) => {
         register: () => {
           if (req.method === "PUT") {
             return {
-              env,
-              url,
-              isProduction,
               deployId: req.query[queryKeys.DeployId]?.toString(),
             };
           }
@@ -53,13 +44,10 @@ export const serve: ServeHandler = (nameOrInngest, fns, opts) => {
         view: () => {
           if (req.method === "GET") {
             return {
-              env,
-              url,
               isIntrospection: Object.hasOwnProperty.call(
                 req.query,
                 queryKeys.Introspect
               ),
-              isProduction,
             };
           }
         },