+++ src/helpers/env.ts
@@ -4,6 +4,7 @@
 // string in order to read variables.
 
 import { envKeys } from "./consts";
+import { stringifyUnknown } from "./strings";
 
 /**
  * devServerHost returns the dev server host by searching for the INNGEST_DEVSERVER_URL
@@ -31,28 +32,68 @@ export const devServerHost = (): string | undefined => {
   return values.find((a) => !!a);
 };
 
-export const isProd = (): boolean => {
-  const values = [
-    processEnv("NODE_ENV"),
-    processEnv("VERCEL_ENV"),
-    processEnv("CONTEXT"),
-  ];
+const prodCheckFns = (<
+  T extends Record<
+    string,
+    (actual: string | undefined, expected: string | undefined) => boolean
+  >
+>(
+  checks: T
+): T => checks)({
+  equals: (actual, expected) => actual === expected,
+  "starts with": (actual, expected) =>
+    expected ? actual?.startsWith(expected) ?? false : false,
+  "is truthy": (actual) => Boolean(actual),
+});
 
-  return values.includes("production");
-};
+const prodChecks: [
+  key: string,
+  customCheck: keyof typeof prodCheckFns,
+  value?: string
+][] = [
+  ["CF_PAGES", "equals", "1"],
+  ["CONTEXT", "starts with", "prod"],
+  ["ENVIRONMENT", "starts with", "prod"],
+  ["NODE_ENV", "starts with", "prod"],
+  ["VERCEL_ENV", "starts with", "prod"],
+  ["DENO_DEPLOYMENT_ID", "is truthy"],
+];
 
-const hasProcessEnv = (): boolean => {
-  return typeof process !== "undefined" && "env" in process;
+/**
+ * Returns `true` if we believe the current environment is production based on
+ * either passed environment variables or `process.env`.
+ */
+export const isProd = (
+  /**
+   * The optional environment variables to use instead of `process.env`.
+   */
+  env: Record<string, unknown> = allProcessEnv()
+): boolean => {
+  return prodChecks.some(([key, checkKey, expected]) => {
+    return prodCheckFns[checkKey](stringifyUnknown(env[key]), expected);
+  });
 };
 
 export const processEnv = (key: string): string | undefined => {
   return allProcessEnv()[key];
 };
 
+declare const Deno: {
+  env: { toObject: () => Record<string, string | undefined> };
+};
+
 export const allProcessEnv = (): Record<string, string | undefined> => {
-  if (hasProcessEnv()) {
+  try {
     // eslint-disable-next-line @inngest/process-warn
     return process.env;
+  } catch (_err) {
+    // noop
+  }
+
+  try {
+    return Deno.env.toObject();
+  } catch (_err) {
+    // noop
   }
 
   return {};